refactor: 重构控制器并提取重复的认证与工具方法
- 在 SpaController 中合并空值检查条件
- 在 AccountPreferencesController 中提取认证检查和工具方法以消除重复代码
- 在 AccountProfileController 中提取认证检查和工具方法以消除重复代码
- 在 HybridRagRetrievalService 中使用 Lombok Getter 替换手写 getter 方法
- 在 WebContentFetchService 中提取 HostAddressResolver 接口以提升可测试性
- 在 LlmImageUploadService 中提取路径标准化逻辑到独立方法
- 修复测试中的类型安全警告和模拟返回值

+++ src/test/java/com/example/EnterpriseRagCommunity/service/monitor/AdminLlmLoadTestServiceCoverageSprintTest.java
@@ -79,7 +79,7 @@ class AdminLlmLoadTestServiceCoverageSprintTest {
     @Test
     void maybeRecomputePaths_should_cover_force_recompute_delta_and_model_fallback_branches() throws Exception {
         TokenCountService tokenCountService = mock(TokenCountService.class);
-        when(tokenCountService.countChatMessagesTokens(any())).thenReturn(5, null);
+        when(tokenCountService.countChatMessagesTokens(any())).thenReturn(5, (Integer) null);
         when(tokenCountService.countTextTokens(anyString())).thenReturn(7);
 
         AdminLlmLoadTestService svc = newService(tokenCountService, mock(LlmQueueMonitorService.class));
